Unittesting for home vieew
add a unit testing project to test the home view if it route correctly.
the test for flight view is in complete since there are no connection
string to the db , so it always return a fail as expected. to unit test
a database need to add a database unit testing , not required.